+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Report exceptions via the `Rails.error.report` interface. ([@palkan][])
+
 ## 1.3.5 (2023-01-04)
 
 - Make misconfiguration error more informative. ([@palkan][])